Refactored wildfire exemplar as package
