+++ wildfire/Canvas_Grid_Visualization.py
@@ -4,8 +4,8 @@ Modular Canvas Rendering
 
 Module for visualizing model objects in grid cells.
 """
-from agents import UAV
-from common_fixed_variables import UAV_OBSERVATION_RADIUS
+from .agents import UAV
+from .common_fixed_variables import UAV_OBSERVATION_RADIUS
 from collections import defaultdict
 
 from mesa.visualization.ModularVisualization import VisualizationElement
