+++ wildfire/__init__.py
@@ -1,0 +1,21 @@
+from .common_fixed_variables import (FIXED_WIND,
+                                     ACTIVATE_SMOKE,
+                                     ACTIVATE_WIND,
+                                     BATCH_SIZE,
+                                     WIDTH,
+                                     HEIGHT,
+                                     BURNING_RATE,
+                                     FIRE_SPREAD_SPEED,
+                                     FUEL_UPPER_LIMIT,
+                                     FUEL_BOTTOM_LIMIT,
+                                     DENSITY_PROB,
+                                     WIND_DIRECTION,
+                                     MU,
+                                     SMOKE_PRE_DISPELLING_COUNTER,
+                                     NUM_AGENTS,
+                                     N_ACTIONS,
+                                     UAV_OBSERVATION_RADIUS,
+                                     N_OBSERVATIONS,
+                                     SECURITY_DISTANCE)
+from .main import main as run_wildfire
+
