+++ wildfire/common_fixed_variables.py
